+++ HISTORY.rst
@@ -1,6 +1,10 @@
 History
 =======
 
+* XML catalog to resolve system ids of type URL
+  [https://github.com/scieloorg/packtools/issues/110].
+
+
 1.3.3 (2017-02-16)
 ------------------